feat: refactor structure of project to use Clean Architecture Pattern
