--- Models/AppSettings.cs
@@ -1,7 +1,0 @@
-﻿namespace CodeChallenge.Models
-{
-    public class AppSettings
-    {
-        public string LocalDBCS { get; set; }
-    }
-}
