--- Startup.cs
@@ -1,47 +1,0 @@
-﻿using CodeChallenge.DAL;
-using CodeChallenge.DAL.Interfaces;
-using CodeChallenge.Models;
-using CodeChallenge.Services.Registrations;
-using Microsoft.EntityFrameworkCore;
-
-namespace CodeChallenge
-{
-    public class Startup
-    {
-        public Startup(IConfiguration configuration, IHostEnvironment env)
-        {
-            Configuration = configuration;
-            Env = env;
-        }
-
-        public IConfiguration Configuration { get; }
-        public IHostEnvironment Env { get; }
-
-        // This method gets called by the runtime. Use this method to add services to the container.
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.Configure<AppSettings>(Configuration);
-            services.AddDbContext<IDbContext, CodeChallengeContext>(options => options.UseSqlServer(Configuration["LocalDBCS"]));
-            services.RegisterWorkerServices();
-            services.AddControllers();
-            services.AddEndpointsApiExplorer();
-            services.AddSwaggerGen();
-        }
-
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
-        {
-            app.UseSwagger();
-            app.UseSwaggerUI();
-            app.UseHttpsRedirection();
-            app.UseRouting();
-            app.UseAuthentication();
-            app.UseAuthorization();
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllers();
-            });
-        }
-    }
-}
