--- UnitsOfWork/DBWorkers/MovieWorkers/GetMovieByDirectorName_DBUOW.cs
@@ -1,42 +1,0 @@
-﻿using CodeChallenge.DAL;
-using CodeChallenge.DAL.Interfaces;
-using CodeChallenge.Models.DTO;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace CodeChallenge.UnitsOfWork.DBWorkers.Movie
-{
-    public class GetMovieByDirectorName_DBUOW
-    {
-        private readonly ILogger<GetMovieByDirectorName_DBUOW> _logger;
-        private readonly IServiceProvider _serviceProvider;
-
-        public GetMovieByDirectorName_DBUOW(
-            ILogger<GetMovieByDirectorName_DBUOW> logger,
-            IServiceProvider serviceProvider)
-        {
-            _logger = logger;
-            _serviceProvider = serviceProvider;
-        }
-
-        public async Task<IEnumerable<Movie_DTO>> ExecuteAsync(string directorName)
-        {
-            var source = this.GetType().Name;
-            _logger.LogInformation($"Starting {source}");
-
-            try
-            {
-                // Create scope
-                using var scope = _serviceProvider.CreateScope();
-                var context = scope.ServiceProvider.GetService<IDbContext>();
-                IMovieRepository movieRepository = new MovieRepository(context);
-
-                return await movieRepository.GetMoviesByDirectorName(directorName);
-            }
-            catch (Exception ex)
-            {
-                throw;
-            }
-        }
-    }
-}
