+++ src/RWE.App.Api/AppSettings.cs
@@ -1,0 +1,7 @@
+﻿namespace RWE.App.Api.Models
+{
+    public class AppSettings
+    {
+        public string LocalDBCS { get; set; }
+    }
+}
