+++ src/RWE.App.Api/Dockerfile
@@ -1,0 +1,22 @@
+#See https://aka.ms/containerfastmode to understand how Visual Studio uses this Dockerfile to build your images for faster debugging.
+
+FROM mcr.microsoft.com/dotnet/aspnet:6.0 AS base
+WORKDIR /app
+EXPOSE 80
+EXPOSE 443
+
+FROM mcr.microsoft.com/dotnet/sdk:6.0 AS build
+WORKDIR /src
+COPY ["src/RWE.App.Api/RWE.App.Api.csproj", "src/RWE.App.Api/"]
+RUN dotnet restore "src/RWE.App.Api/RWE.App.Api.csproj"
+COPY . .
+WORKDIR "/src/src/RWE.App.Api"
+RUN dotnet build "RWE.App.Api.csproj" -c Release -o /app/build
+
+FROM build AS publish
+RUN dotnet publish "RWE.App.Api.csproj" -c Release -o /app/publish /p:UseAppHost=false
+
+FROM base AS final
+WORKDIR /app
+COPY --from=publish /app/publish .
+ENTRYPOINT ["dotnet", "RWE.App.Api.dll"]
