+++ src/RWE.App.Core/Entities/Director.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace RWE.App.Core.Entities;
+
+public partial class Director
+{
+    public Guid Uuid { get; set; }
+
+    public string Name { get; set; }
+
+    public DateTime Birthdate { get; set; }
+
+    public virtual Movie Movie { get; set; }
+}
