+++ src/RWE.App.Core/Interfaces/IUnitOfWork.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace RWE.App.Core.Interfaces
+{
+    public interface IUnitOfWork
+    {
+        IMovieRepository MovieRepository { get; }
+        IDirectorRepository DirectorRepository { get; }
+        Task SaveChangeAsync();
+    }
+}
