+++ src/RWE.App.Infrastructure/RWE.App.Infrastructure.csproj
@@ -1,0 +1,35 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Remove="UnitsOfWork\ControllerWorkers\**" />
+    <Compile Remove="UnitsOfWork\DBWorkers\**" />
+    <EmbeddedResource Remove="UnitsOfWork\ControllerWorkers\**" />
+    <EmbeddedResource Remove="UnitsOfWork\DBWorkers\**" />
+    <None Remove="UnitsOfWork\ControllerWorkers\**" />
+    <None Remove="UnitsOfWork\DBWorkers\**" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Remove="Class1.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="7.0.4" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="7.0.4">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="7.0.4" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\RWE.App.Core\RWE.App.Core.csproj" />
+  </ItemGroup>
+
+</Project>
